+++ src/types/TableFieldStrings.ts
@@ -1,0 +1,1 @@
+export type TableFieldStrings<T> = (keyof T)[]